FEATURE: update sdk version to target latest stable .net core SDK FEATURE: use default empty overload in sample FEATURE: rename GetLogEventLevel to GetLogLevel
