+++ global.json
@@ -1,5 +1,5 @@
 {
   "sdk": {
-    "version": "2.2.105"
+    "version": "2.2.402"
   }
 }
