+++ samples/InlineInitializationSample/Startup.cs
@@ -48,14 +48,7 @@ namespace InlineInitializationSample
             // Write streamlined request completion events, instead of the more verbose ones from the framework.
             // To use the default framework request logging instead, remove this line and set the "Microsoft"
             // level in appsettings.json to "Information".
-            app.UseSerilogRequestLogging(opts =>
-            {
-                opts.GetLogEventLevel = statusCode =>
-                    statusCode == HttpStatusCode.InternalServerError
-                        ? LogEventLevel.Error
-                        : LogEventLevel.Information;
-                // OR opts.GetLogEventLevel = _ => LogEventLevel.Information;
-            });
+            app.UseSerilogRequestLogging();
 
             app.UseStaticFiles();
             app.UseCookiePolicy();
